// src/interfaces/enums.ts
export enum ActionTypes {
    PLAY_PAUSE = 'com.pear.desktop.streamdeck.play-pause',
    NEXT_TRACK = 'com.pear.desktop.streamdeck.next',
    PREV_TRACK = 'com.pear.desktop.streamdeck.prev',
    VOLUME_DOWN = 'com.pear.desktop.streamdeck.volume-down',
    VOLUME_UP = 'com.pear.desktop.streamdeck.volume-up',
    VOLUME_MUTE = 'com.pear.desktop.streamdeck.mute',
    LIKE_TRACK = 'com.pear.desktop.streamdeck.like',
    DISLIKE_TRACK = 'com.pear.desktop.streamdeck.dislike',
    SONG_INFO = 'com.pear.desktop.streamdeck.song-info',
    SHUFFLE = 'com.pear.desktop.streamdeck.shuffle',
    REPEAT = 'com.pear.desktop.streamdeck.repeat',
    PLAY_PLAYLIST = 'com.pear.desktop.streamdeck.play-playlist'
}